requirements.txt
langchain
pinecone-client
langchain-community
sentence-transformers
Streamlit
PyPDF2
python-docx